refactor: WorksPageをリファクタ
