--- src/components/WorksPage/index.ts
@@ -1,2 +1,0 @@
-export * from './Achievements'
-export * from './RecentPosts'
