+++ src/components/page/Works/index.ts
@@ -1,0 +1,1 @@
+export * from './Works.page'
